[管理側]Station Controllerの名前空間をAdmin::Stationにする
