+++ app/views/admin/stations/_form.html.slim
@@ -1,0 +1,3 @@
+= form_with model: [:admin, station] do |form|
+  = form.text_field :name
+  = form.submit
