+++ app/views/admin/stations/edit.html.slim
@@ -1,0 +1,2 @@
+h3 #{@station.name}の駅名変更
+= render 'form', station: @station
