+++ app/views/admin/stations/index.html.slim
@@ -1,0 +1,5 @@
+h1 トップページ
+= link_to '駅登録', new_admin_station_path
+br
+- @stations.each do |station|
+  = link_to station.name, admin_station_path(station)
