+++ app/views/admin/stations/show.html.slim
@@ -1,0 +1,3 @@
+h1 = @station.name
+= link_to '駅名変更', edit_admin_station_path
+= link_to '駅削除', admin_station_path(@station), method: :delete
