--- app/views/stations/show.html.slim
@@ -1,3 +1,0 @@
-h1 = @station.name
-= link_to '駅名変更', edit_station_path
-= link_to '駅削除', "/stations/#{@station.id}", method: :delete
